+++ kibot/PyPDF2/__init__.py
@@ -1,0 +1,2 @@
+from .merger import PdfFileMerger
+__all__ = ["pdf", "PdfFileMerger"]